minor warning fixes

+++ src/utils/string_conv.h
@@ -27,7 +27,7 @@
 extern int char2int(char input);
 extern size_t hex2bin(const char* src, char* target);
 extern char getIcaoChar (unsigned char c);
-extern std::string binary2hex(const unsigned char* src, size_t length);
+extern std::string binary2hex(const unsigned char* src, unsigned int length);
 
 extern std::vector<std::string> &split(const std::string &s, char delim, std::vector<std::string> &elems);
 //extern std::vector<std::string> split(const std::string &s, char delim);
